+++ samples/testing-frameworks/appium/client-side/csharp/biometric-sample-test/BitbarBiometricSample/BitbarBiometricSample.csproj
@@ -15,6 +15,7 @@
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
     <PackageReference Include="coverlet.collector" Version="3.1.2" />
+    <PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
   </ItemGroup>
 
 </Project>
