+++ samples/testing-frameworks/appium/client-side/csharp/biometric-sample-test/BitbarBiometricSample/TestBiometricsIOS.cs
@@ -1,32 +1,24 @@
 ﻿using OpenQA.Selenium.Appium.iOS;
+using OpenQA.Selenium.Support.UI;
 
 namespace BitbarBiometricSample
 {
     [TestFixture]
-    public class TestBiometricsIOS
+    public class TestBiometricsIOS : BaseBiometricTest
     {
-        private readonly String SCREENSHOT_DIRECTORY = Directory.GetParent(Directory.GetCurrentDirectory()).Parent.Parent.FullName + "/screenshots/";
-        private RemoteWebDriver driver;
-
         [OneTimeSetUp]
         public void InitDriver()
         {
-            const String AppiumHubURL = "https://appium.bitbar.com/wd/hub";
-
-            AppiumOptions capabilities = new();
+            //Platform dependent
             capabilities.AddAdditionalCapability("platformName", "iOS");
             capabilities.AddAdditionalCapability("deviceName", "iPhone device");
             capabilities.AddAdditionalCapability("automationName", "XCUITest");
-
             //Customizable
             capabilities.AddAdditionalCapability("bitbar_project", "C# Appium Automated Test");
             capabilities.AddAdditionalCapability("bitbar_testrun", "iOS Run");
             capabilities.AddAdditionalCapability("bitbar_device", "Apple iPhone 11");
-            //Required
-            capabilities.AddAdditionalCapability("bitbar_apiKey", "<YOUR_BITBAR_API_KEY>");
+            //App ID
             capabilities.AddAdditionalCapability("bitbar_app", "<APP_ID>");
-            //Biometrics capability
-            capabilities.AddAdditionalCapability("biometricInstrumentation", true);
 
             Console.WriteLine("Sending request to start a session. Waiting for response typically takes 2-3 minutes...");
             driver = new IOSDriver<IOSElement>(new Uri(AppiumHubURL), capabilities, TimeSpan.FromSeconds(300));
@@ -42,41 +34,28 @@ namespace BitbarBiometricSample
         [Test]
         public void Test_Biometric_Authentication()
         {
+            Console.WriteLine("Test execution started:");
+
             //Accept initial biometric dialog
-            Thread.Sleep(10000);
-            driver.SwitchTo().Alert().Accept();
+            var wait = new WebDriverWait(driver, TimeSpan.FromSeconds(10)).Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementExists(By.XPath("//XCUIElementTypeAlert[@name=\"Biometric Authentication\"]")));
+            driver.ExecuteScript("mobile: alert", new Dictionary<string, string> { { "action", "accept" }, { "buttonLabel", "OK" } });
 
             TakeScreenshot("HomeScreen");
             var buttonBiometrics = driver.FindElement(By.XPath("(//*[@name=\"Biometric authentication\"])[3]"));
             buttonBiometrics.Click();
 
             var textSensorType = driver.FindElement(MobileBy.AccessibilityId("Sensor type value"));
-            Assert.That(textSensorType.Text.ToUpper, Is.EqualTo("TOUCHID"));
+            Assert.That(textSensorType.Text.ToUpper, Is.EqualTo("TOUCHID").Or.EqualTo("FACEID"));
 
             var buttonAskBiometrics = driver.FindElement(By.XPath("//XCUIElementTypeOther[@name=\"Ask biometric authentication\"]"));
             buttonAskBiometrics.Click();
             TakeScreenshot("Dialog");
 
-            driver.SwitchTo().Alert().Accept();
+            driver.ExecuteScript("mobile: alert", new Dictionary<string, string> { { "action", "accept" }, { "buttonLabel", "Pass" } });
 
             var textResult = driver.FindElement(MobileBy.AccessibilityId("Authentication status value"));
             TakeScreenshot("Result");
             Assert.That(textResult.Text.ToUpper, Is.EqualTo("SUCCEEDED"));
-        }
-
-        public void TakeScreenshot(String fileName)
-        {
-            try
-            {
-                Console.WriteLine("Taking screenshot: " + fileName + ".png");
-                var filePath = SCREENSHOT_DIRECTORY + fileName + ".png";
-                var screenshot = driver.GetScreenshot();
-                screenshot.SaveAsFile(filePath, ScreenshotImageFormat.Png);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Failed to take a screenshot: " + e.Message);
-            }
         }
     }
 }
