+++ samples/testing-frameworks/appium/client-side/csharp/biometric-sample-test/BitbarBiometricSample/Usings.cs
@@ -2,4 +2,3 @@ global using NUnit.Framework;
 global using OpenQA.Selenium.Remote;
 global using OpenQA.Selenium;
 global using OpenQA.Selenium.Appium;
-global using OpenQA.Selenium.Appium.Android;
