+++ samples/testing-frameworks/appium/client-side/csharp/sample-test/Test123/TestdroidAndroidSample.csproj
@@ -30,10 +30,6 @@
     <WarningLevel>4</WarningLevel>
     <ConsolePause>false</ConsolePause>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\x64\Debug\</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
